+++ lib/src/inputs/src/pickers/general/file.dart
@@ -1,81 +1,5 @@
 part of '../../../inputs.dart';
 
-class ThemedFileInput extends StatelessWidget {
-  final String? labelText;
-  final Widget? label;
-  final String? value;
-  final void Function(String, List<int>)? onChanged;
-  final FileType acceptedTypes;
-  final bool disabled;
-  final List<String> errors;
-  final bool hideDetails;
-  final bool isRequired;
-  final Widget? customChild;
-  final Widget? customWidget;
-
-  /// [ThemedFileInput] is the input for file selection.
-  /// It uses [ThemedTextInput] as the base.
-  @Deprecated('Use `ThemedFilePicker` instead')
-  const ThemedFileInput({
-    super.key,
-
-    /// [labelText] is the label of the input. Avoid using this if you are using [label] instead.
-    this.labelText,
-
-    /// [label] is the label of the input. Avoid using this if you are using [labelText] instead.
-    this.label,
-
-    /// [value] is the value of the input.
-    this.value,
-
-    /// [onChanged] is the callback when the input value changes.
-    /// The first parameter is the base64 of the file.
-    /// The second parameter is the byte array of the file.
-    this.onChanged,
-
-    /// [disabled] is the flag to disable the input.
-    this.disabled = false,
-
-    /// [errors] is the list of errors to be displayed.
-    this.errors = const [],
-
-    /// [hideDetails] is the flag to hide the details of the input.
-    this.hideDetails = false,
-
-    /// [isRequired] is the flag to mark the input as required.
-    this.isRequired = false,
-
-    /// [acceptedTypes] is the type of files that can be selected.
-    this.acceptedTypes = FileType.any,
-
-    /// [customChild] is the custom widget to be displayed.
-    /// Replaces the [ThemedTextInput] widget.
-    this.customChild,
-
-    /// [customWidget] is the custom widget to be displayed.
-    /// Replaces the [ThemedTextInput] widget.
-    @Deprecated('Use `customChild` instead') this.customWidget,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return ThemedFilePicker(
-      labelText: labelText,
-      label: label,
-      value: value,
-      onChanged: onChanged,
-      acceptedTypes: acceptedTypes,
-      disabled: disabled,
-      errors: errors,
-      hideDetails: hideDetails,
-      isRequired: isRequired,
-      customChild: customChild,
-      // ignore: deprecated_member_use_from_same_package
-      customWidget: customWidget,
-    );
-  }
-}
-
 class ThemedFilePicker extends StatefulWidget {
   /// [labelText] is the label of the input. Avoid using this if you are using [label] instead.
   final String? labelText;
@@ -109,10 +33,6 @@ class ThemedFilePicker extends StatefulWidget {
   /// [customChild] is the custom widget to be displayed.
   /// Replaces the [ThemedTextInput] widget.
   final Widget? customChild;
-
-  /// [customWidget] is the custom widget to be displayed.
-  /// Replaces the [ThemedTextInput] widget.
-  final Widget? customWidget;
 
   /// [hoverColor] is the hover color of the input. Only will affect when [customChild] is submitted.
   /// By default, it will use `Colors.transparent`.
@@ -154,7 +74,6 @@ class ThemedFilePicker extends StatefulWidget {
     this.isRequired = false,
     this.acceptedTypes = FileType.any,
     this.customChild,
-    @Deprecated('Use `customChild` instead') this.customWidget,
     this.hoverColor = Colors.transparent,
     this.focusColor = Colors.transparent,
     this.splashColor = Colors.transparent,
@@ -170,7 +89,7 @@ class ThemedFilePicker extends StatefulWidget {
 
 class _ThemedFilePickerState extends State<ThemedFilePicker> with SingleTickerProviderStateMixin {
   final TextEditingController _controller = TextEditingController();
-  Widget? get customChild => widget.customChild ?? widget.customWidget;
+  Widget? get customChild => widget.customChild;
 
   String _value = "";